backend/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-totp==0.1.4
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
jq==1.8.0
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
